Move setup to separate file
