+++ src/games/colonizers/Buildings.tsx
@@ -2,12 +2,10 @@ import React from 'react';
 import { IG, isValidBuildingPosition, isAnyOwnRoadConnected, Building } from './game';
 import { IGameCtx } from '@freeboardgame.org/boardgame.io/core';
 import { dirToAngle, getTilePos } from './board';
-import { IGameArgs } from '../../App/Game/GameBoardWrapper';
 import { Phase } from './phase';
 import css from './Building.css';
 
 import grey from '@material-ui/core/colors/grey';
-import { isLocalGame } from '../../common/gameMode';
 
 interface IBuildingsProps {
   G: IG;
@@ -15,7 +13,6 @@ interface IBuildingsProps {
   playerColors: string[];
   selectedRecipe: number;
   selectedBuilding: number;
-  gameArgs: IGameArgs;
   playerID: string;
   size: number;
   click: (index: number) => void;
@@ -31,9 +28,8 @@ export class Buildings extends React.Component<IBuildingsProps, {}> {
               building.owner !== null ||
               (isValidBuildingPosition(this.props.G, building.index) &&
                 ((this.props.selectedRecipe === Building.Settlement &&
-                  isAnyOwnRoadConnected(this.props.G, this.props.ctx.currentPlayer, building.index)) ||
-                  (this.props.ctx.phase === Phase.Place &&
-                    (isLocalGame(this.props.gameArgs) || this.props.ctx.currentPlayer === this.props.playerID)))),
+                  isAnyOwnRoadConnected(this.props.G, this.props.playerID, building.index)) ||
+                  (this.props.ctx.phase === Phase.Place && this.props.ctx.currentPlayer === this.props.playerID))),
           )
           .map(building => {
             const angle = dirToAngle(building.tileRefs[0].dir);
@@ -41,7 +37,7 @@ export class Buildings extends React.Component<IBuildingsProps, {}> {
             const selected = this.props.selectedBuilding === building.index;
             let fill = building.owner === null ? grey[900] : this.props.playerColors[building.owner as any];
             if (selected) {
-              fill = this.props.playerColors[this.props.ctx.currentPlayer as any];
+              fill = this.props.playerColors[this.props.playerID as any];
             }
 
             return (
